[REFACTOR] ControllerTestSupport 패키지 이동
