+++ src/test/java/kr/hhplus/be/ecommerce/interfaces/balance/BalanceControllerTest.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.ecommerce.interfaces.balance;
 
 import kr.hhplus.be.ecommerce.application.balance.BalanceResult;
-import kr.hhplus.be.ecommerce.interfaces.ControllerTestSupport;
+import kr.hhplus.be.ecommerce.support.ControllerTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
